Working
This new design is the same as the old evaluator. The old approach wanted none of the chesspieces to own their own move logic. But as the evaluator began having more reusable lines, it became inevitable to take this approach.

Now evaluator just filters out the legal moves based on the game state.